Use labels to ensure node initiate encrypting its routes.
When a node connects to a cluster, it is possible that
kube master might see the new node and start encrypting connection before
the node has a change to get list of nodes. This change ensures that nodes
always get a chance to read the node list first. A node will first read the
list. If it does not have it's "net.beta.appscode.com/vpn": "true", it will
apply the label. Then update ipsec daemon. At this point other nodes
(including master) will start encrypting their end of the connection.

To work properly, /etc/ipsec.conf needs to be persisted across restarts
for a given node.